create api invoice/create

+++ invoiceApplication/src/main/java/application/repository/InvoiceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import application.model.Invoice;
 
-public interface InvoiceRepository extends JpaRepository<Invoice, Integer>  {
+public interface InvoiceRepository extends JpaRepository<Invoice, Long>  {
 
 	List<Invoice> findByInvoiceCode(String invoiceCode);
